+++ src/FamilyHubs.SharedKernel.Razor/Dashboard/LargeSetLinkPagination.cs
@@ -1,4 +1,5 @@
 ﻿using FamilyHubs.SharedKernel.Razor.Pagination;
+using System.Web;
 
 namespace FamilyHubs.SharedKernel.Razor.Dashboard;
 
@@ -15,17 +16,38 @@ public class LargeSetLinkPagination<TColumn> : LargeSetPagination, ILinkPaginati
     private readonly string _dashboardPath;
     private readonly TColumn _column;
     private readonly SortOrder _sort;
+    private readonly string? _extraQueryParams;
 
-    public LargeSetLinkPagination(string dashboardPath, int totalPages, int currentPage, TColumn column, SortOrder sort)
+    public LargeSetLinkPagination(
+        string dashboardPath, int totalPages, int currentPage,
+        TColumn column, SortOrder sort,
+        //todo: or IQueryCollection? or extra constructor
+        IReadOnlyDictionary<string, string> extraQueryParams)
+
+        : this(dashboardPath, totalPages, currentPage, column, sort,
+            string.Join('&', extraQueryParams.Select(kvp => $"{HttpUtility.UrlEncode(kvp.Key)}={HttpUtility.UrlEncode(kvp.Value)}")))
+    {
+    }
+
+    public LargeSetLinkPagination(
+        string dashboardPath,
+        int totalPages,
+        int currentPage,
+        TColumn column,
+        SortOrder sort,
+        string? extraQueryParams = null)
+
         : base(totalPages, currentPage)
     {
         _dashboardPath = dashboardPath;
         _column = column;
         _sort = sort;
+        _extraQueryParams = extraQueryParams;
     }
 
     public string GetUrl(int page)
     {
-        return $"{_dashboardPath}?columnName={_column}&sort={_sort}&currentPage={page}";
+        var url = $"{_dashboardPath}?columnName={_column}&sort={_sort}&currentPage={page}";
+        return _extraQueryParams == null ? url : $"{url}&{_extraQueryParams}";
     }
 }